Add find_best_path function with costs for cells
New path finding function:

	find_best_path(start_cell, target_cell, obstacles=None,
		allow_obstacles=False, randomize=True, cost_func=None,
		set_path_cost=None)

The find_path function has a distance from the start_cell as the weight.

This new find_best_path function is somewhat similar, but it gets the
cost function for a cell and finds the best path by accumulating costs
along the path. For each neighbour cell it stores the parent cell (to
reconstruct the path backwards from the target cell when finished) and
the accumulated cost. If there is another sub-path to the already visited
cell, then its parent and the cost are replaced for this cell.

The algorythm uses A Star to determine the next cell to be processed.
The weight is the accumulated cost plus the distance to the target_cell.
The algorythm is finished when the target cell is reached. Then the path
is reconstructed and returned (list of cells, including target_cell and
excluding start_cell). If no path is found, None is returned.
If start_cell equals to target_cell then empty path is returned.

If the set_path_cost array of 1 element is given as an argument, then its
first element is set to the accumulated target_cell cost. Only when the
path exists and it is not empty. The caller is responsible to set the
default path_cost for all other cases.

The cost function has this interface:

cost_func(neigh, cell, visited_cells, start_cell, target_cell, obstacles)

When cost_func returns None this means the cell is not accessible.
Otherwise the returned cost is used to get the accumulated cost along the
path that is set for the processed cell.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>