BarrelPuzzle: Implement new generation algorithm and update levels
The old generation algorithm is fast, but not very good. It places the
barrels on plates and tries to pull them back until possible.

The new generation algorithm ("use_ng" in config) is different.

For each iteration (up to 2000), try to generate good level:

	* initialize the whole room to walls
	* set random plates, barrels and starting char cell in the area
	* iteratively find paths from barrel to plate and remove walls

For each generate_char_and_barrel_paths_to_plates iteration (up to 4000):

	* use find_best_path to calculate paths from each unplaced barrel
	  to its plate and from char to the start of the barrel path
	* apply the char path and start of the barrel path until the char
	  changes direction

This may take time, but it generates good levels.

What is currently missing is the optimization stage to find a solution
for the generated level and set unused floor cells to walls.

Add a new easy timer-based level with 9 rooms, with 2 barrels per room.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>