Add support for special_cell_types in load_map and update puzzles
Now each puzzle can define load_map_special_cell_types. By default it is
{CELL_SPECIAL0: 'str'} that is good for AtomixPuzzle. The default is set
to guide other puzzles. The format of this dict is cell_type: 'str',
'int', ints', and possibly other value types in the future (like 'cell'
and 'cells' that could be a pair of int). The value type determines how
the special line in map file that comes after the map itself is parsed.
If it is 'str' then it is left as is, if 'int', then it is converted to
int, and 'ints' are split by whitespace first.

A meaningful error is thrown on special cell value parsing errors.

Updated AtomixPuzzle and GatePuzzle correspondingly.

In GatePuzzle (and in future SwitchPuzzle), load_map_special_cell_types
is {CELL_PLATE: 'ints'}.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>