CellActor: Support cell and opacity in draw()
If optional cell is given the actor is (temporarily) drawn on this cell.
This does not affect the persistent actor location (that is restored).

If optional opacity is given the the actor is drawn using it. This does
not affect the persistent actor opacity (that is restored).

Simplify calls in draw_map() and draw_status() to use this new feature.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>