Redesign status messages with new module and API
Move all status message logic into new statusmessage.py and add internal
class StatusMessage and this functional API:

	set_status_message msg source priority duration
	draw_status_message
	reset_status_messages

	get_fade_text_factor curr_time end_time fade_duration rest_duration

The messages are unique using source+priority pair. The default source is
"main", but the puzzles are supposed to pass self as the source.

The priority is currently the number from 0 to 10. The priority 0 is
special in that the status messages with it are displayed and do not
yield to messages with other priorities. The priority 10 is special in
that the status messages with other priorities always are displayed and
do not yield.

The default priority is currently 5.

Each status message can be unlimited (will be displayed until replaced or
removed) when duration is None or limited in time if duration is positive
integer. The duration indicates the total time when the status message is
going to be displayed. But it may be divided, i.e. different messages are
displayed one after another, possibly in a loop. But the total time of
each message will be the specified duration.

If msg given in set_status_message is None, this means the corresponding
unique message given by source+priority (if any) is removed.

The messages (if several) are selected by priority and the number of
times the message was displayed already. The minimal display time is
BEAT_TIME that is currently 5 seconds.

The messages are drawn using fade out effect (via get_fade_text_factor).

This is a pretty flexible mechanism. But it may be improved in the future
if needed.

Also, show "You lost" status message.

Also, slightly improve keystroke 'o' displaying "Hello, world" for debug.

Change all puzzles accordingly to the new API.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>