BarrelPuzzle: Implement more experimental solution code
There is a new experimental try to optimize solution using KP_Enter key.

There is now an optional prepare-solution stage for max of 8 seconds.
This stage may be toggled on and off by pressing 'd'.

The unaccessible cells are now colored with Red during solution. This is
likely to be changed.

There will be more experiments in the near future in hope to make the
solver faster and more useful.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>