Add new optional params to debug_map
New param "char_cell" is an alternative way to char.c (in cases when
it is None, i.e. during level generation).

New param "cell_chars" is dict with cell-to-ch to be used to represent
given cells. For example, you may use this param to add some path to the
map when the path cells are represented by '⨯' or by numbers.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>