Add "state" concept to find_best_path
The BarrelPuzzle when finding solution does not require additional state
other than the player char cell. However other puzzles (like
MinotaurPuzzle) may require extra state (position of minotaur).

Support both state-less and state-full modes in this function.

New param "state_func" if given is used to calculate the state of the
position similarly to cost. The params of this state_func is:
	new_char_cell, old_char_cell, old_state

It should return new_state or None if the position becomes forbidden and
should be excluded from the solution path.

Currently BarrelPuzzle uses cost_func, but not state_func. And
MinotaurPuzzle will use state_func, but not cost_func. And in both cases
the A* algorithm uses distance to the target_cell. However, both
cost_func and state_func may be used in theory.

The internal data structs visited_cells, processed_cells and
unprocessed_cells now use cell_state if state_func is given (i.e. tuple
(cell, state)) and just cell when it is not given as their unique key.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>