MinotaurPuzzle: Change calculate_minotaur_move to work with any actor cells
Make it possible not to hardcode the minotaur and char cells when
calculating new minotaur move by passing char_cell and minotaur_cell,
although still use char.c and self.minotaur.c if these are not given.

This will be used for generation when there are no char and minotaur
positions finalized in the room yet.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>