MinotaurPuzzle: In generator check that all shortest paths lead to lose
It is not enough that any random shortest path leads to defeat for the
room setup to qualify. All shortest paths should lead to lose for the
room setup to be considered interesting.

Support config flag "allow_partially_trivial" to preserve the previous
logic. This is needed for one level with nine 3x3 rooms. Without this
flag there is only one setup with truely non-trivial solution that is not
desired. So we want to allow rooms with partially trivial solutions too.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>