Implement Solution centrally for all puzzles that support it
Add new module solution.py with classes SolutionItem and Solution.

Have a global instance solution that may be in one of several states:
	* inactive (no solution)
	* find mode (multi-stage find-solution supported)
	* active (solution is found and non-empty)
	* play mode (solution is being played)

The first KP Enter now switches to find-mode (if the puzzle supports it).

The second KP Enter (if solution is found) switches to play-mode when the
moves are automatically replayed played until solution is active.

The puzzle may return solution as a list of items, each may be:
	* cell to press
	* target cell to move to
	* cells to move as the char path

When target cell is encountered in the solution it is expanded to the
cells to move to using find_path at real time (not at the solution set or
the generation time).

The puzzles should override the prepare_solution method to return find
solution message and find-solution function. This find-solution function
(usually the puzzle method) will be called once the message is drawn. The
find-solution function should return 2 values, potential solution items
or None, and potential new message if another stage is needed (that will
be again called after this new message is drawn, and so on). If there is
no solution found, the find-solution function should return None, None.

The concrete puzzle implementation of this new solution API will come
in separate commits.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>