Add visual effect when pressing cell other than the char cell
This may be seen in the solver of ColorPuzzle (and possibly more puzzles
in the future). When the char cell is pressed it's more or less intuitive
without an extra visual effect. But it's better to add a visual effect
when pressing non-char cells. Do this using the existing cursor image.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>