Add support for button-to-press in "press_cell" solution item
The press_cell concept includes button being pressed. For keyboard press
(using Space) the button is None, for mouse presses and special keys
(Insert, Delete, Home, End, PgUp and PgDown) the button is an integer.

Add support for button in press_cell during the solution.

Use it in RotatepicPuzzle. It is always possible to press only one button
instead of 2 or 3 to bring any cell to the solved position.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>