MinotaurPuzzle: Highly improve generator
Now 8000 random levels are tried with the given size.

Now the goal cell can be any area cell and not necessarily the
right-bottom cell.

Now 25 good positions are generated and the best is taken. Currently the
best is calculated by number of solution moves (the longer the better).
Possibly in the future the minotaur path len will be considered too.

Now the level generation is limited to 7 seconds.

Now the long_generation of puzzle is turned on (flash message while
generating).

Now there is a different approach at removing area walls. About half of
the walls (after placing actors and generating paths from them to goal)
are removed.

Fixed the condition when the portal is placed in the level (when the room
and the area are not the same size and it is possible to place a portal).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>