MinotaurPuzzle: Improve criteria of what is a "good level"
Use weight to determine the level goodness. Now the solution is examined
and the weight is 67% of the char path len in the solution plus 33% of
the minotaur path len in the solution.

Possibly later the level goodness weight will include other criterias and
different factors. But intuitivelly such weight formula should generate
levels with interesting solutions given enough time.

Of course, the current 7 seconds and 25 playable levels to examine at
most is not enough to find brilliant levels. In the future it may be good
to have manual levels generated within long time and lots of tries.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>