Add new actor type "npc" for predefined maps
The concrete puzzle may interpret new actor type "npc" on its own. For
example MinotaurPuzzle will take it as minotaur (single for now, but
possibly multiple minotaurs will be supported per room in the future).

Once the npc char is detected on the map, it will be handled similarly to
other actor chars on the map (as usual, there are 2 variants, on the
floor and on the plate). But unlike other actors (like barrels, enemies,
drops, char), this one will not be created, but supplied in
special_cell_values dict with None value to puzzle's on_load_map method.

In the future, predefined maps with this new actor type will be added.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>