CellActor: Add proper support for hidden and phased state
Now actors may be phased. It is somewhat similar to hidden when the actor
is not drawn at all and should not participate in cell accessibility and
cell occupation (now this is explicit). In the phased state the actor is
drawn slightly visible (opacity 0.4 by default) and is not intended to
participate in cell accessibility (i.e. other actors may pass through
it), but may participate in cell occupation (i.e. the plate is weighted).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>