Puzzle: Introduce is_char_phased and use in puzzles
Add new puzzle callback that returns False by default:
	puzzle.is_char_phased()

It is currently called on enter_room (after on_enter_room) and in
prepare_enter_cell (after on_prepare_enter_cell).

Use this in Atomix, Fifteen, Memory and Rotatepic puzzles.

Now there is no need for these puzzles to listen to update() that is
several times per second. They just implement this boolean callback that
is called when char starts or is moved. And now there is no need to reset
the phased state on finish(). The char state is automatically reset on
every new level.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>