In load_map support None elements in special cell values "int" and "ints"
The None elements should be encoded as "-" in maps.

I.e. special cell line "12 - 3 4 5" is parsed as (12, None, 3, 4, 5).

Usually this should not be used, except for well-defined cases, like for
SwitchBoxPuzzle in the next commit. It's now up to puzzles to decide what
to do with None elements, and whether they are valid.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>