Implement toggle_actor_phased with sound and animation
Now when actor is phased in, sound "switch-on" is played, and the actor
opacity is gradually changed from 1 to 0.4. When actor is phased out,
sound "switch-off" is played, and the actor opacity is gradually changed
from 0.4 to 1.

Currently these config constants are used:
	ACTOR_PHASED_OPACITY = 0.4
	ACTOR_PHASED_DURATION = 0.3

The new sounds are taken from free internet resources and edited in
audacity.

Use this new function in SwitchBoxPuzzle.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>