Allow to specify level id string in config
Currently the level id string is something like "Level: 5.1". Allow to
specify this string in config, like "SwitchBox - level 4" as "level" key
in level config.

In the future, when level collections will be reworked (i.e. collections
will be named instead of hardcoding the major level num), this level id
string may be automatically composed of collection name and the actual
level num in the collection.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>