Redesign the effect of char looking to the left or to the right
Do not have 2 images for char anymore.

Remove historical "stand" and "left" images for char and add one "char"
image (prior "stand").

Use the CellActor flip feature to flip char horisontally when needed.

Add a new feature. Determine the initial char flip direction by the
position of the char (either the left or right half of the room). This is
done in enter_room. Previously the old level or room direction was used.

Determine the direction of the char during the movement just like it was.
I.e. on the left and right movement set the flip direction, and on the up
and down movement do not change it.

Do not pass flip param to the char appearing animation with CELL_START
(in stoneage levels). The implementation of it is broken anyway.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>