Make char, cursor and enemy images theme-able
Create global actors char and cursor without initial image. The images
are now set on theme set. The same with enemies.

Rename char.png to viking.png
Rename cursor.png to cursor-round-yellow.png.

Create symlinks in images/default/:
	char.png -> ../viking.png
	enemy.png -> ../skeleton.png
	cursor.png -> ../cursor-round-yellow.png

The char and the enemy on the corresponding sprites are expected to look
at the right. They will be auto-flipped horizontally when needed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>