Centrally unset prepared solution and cancel playing solution
Add unset_prepared_solution() to unset solution and call it on char move
and press cell.

Add cancel_playing_solution() to cancel ongoing solution and call it on
any key or mouse press.

Have press_cell and press_cell_in_solution. In the first call
unset_prepared_solution, but not in the second (obviously).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>