Improve the logic of level title, name and goal
There are level title and name as well as goal that are shown at the
level start as usual. Improve their defaults and logic.

The level title can be configured in the level config by "title" key. It
is taken as "main-screen" for the main screen. If the level title is not
configured it is taken as "{level-label} %d", i.e. "Level 3.4".

The level name can be configured in the level config by "name" key, If
the level name is not configured, it is taken as "level-3.4-name" iff
there is translation of this string. Otherwise the name is not shown.

The level goal can be configured in the level config by "goal" key.
Otherwise if puzzle.is_goal_to_be_solved() the goal key is taken as
"solve-<canonical_puzzle_name>-puzzle" (i.e. "solve-barrel-puzzle").
Otherwise if puzzle.is_goal_to_kill_enemies() the goal key is taken as
"kill-enemies". Otherwise goal key is taken as "reach-finish".

If the level config contains "time_limit" it's value is appended to the
level goal, like "<level-goal> {in-word} %d {seconds-word}" i.e. "Reach
the finish in 75 seconds".

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>