TrafficPuzzle: Add a new puzzle initially
In Traffic puzzle there are "cars" that can move either horizontally or
vertically. The goal is to move other "cars" to open the way for the
target "cars" to their exit. The target "cars" and their corresponding
exits are color-coded.

We currently use our existing lifts (vertical and horizontal) for cars.
But this should be changed, since a car should not behave like a lift; it
should move freely and not all the way until obstacle. And the real cars
should better be moved on the ground (floor) and not in the void.

In the future possibly multi-size lifts will be supported (larger than
1x1), and possibly some of them could move in all 4 directions or just in
one direction, just like our existing lifts.

This puzzle is similar to OriMaze, but ideally it will be much more
capable, similar to classic Rush Hour puzzle.

Support config values "num_free_cells" and "num_targets".

Support generation of random nonsolvable room.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>