Use only ascii chars in main.py to workaround pgzero problem
It was fixed in pgzero on 2019-05-06 in commit 97c989dda2c7 ("Compile
module source from bytes to avoid encoding issues"), but it looks like
the pip still does not have this fix.

So, use the plain ascii for main.py and encode unicode chars using
"\uxxxx" syntax.

This fixes critical problems of running the game at least on Windows.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>