MinotaurPuzzle: Refactor generate_best_random_setups to use in other puzzles
Move the logic for generation of rooms from MinotaurPuzzle to Puzzle.
This will be used in TrafficPuzzle.

The logic is the puzzle should generate a random setup and tell whether
it is good (then it returns weight of this setup and the setup data) or
bad (then it returns None). Stop when any of the 3 given conditions meet:
1) reached the max number of setups to try, 2) reached the max number of
good setups to try, 3) reached the time limit.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>