TrafficPuzzle: Add experimental new code for CarPark and Car
Add classes Car and CarPark.

Add generation of levels that respects the config keys "num_free_cells",
"num_targets" and "size". There is no scramble yet. The lifts are still
used instead of cars. And the generation is not really good yet. And
there is currently recursion limit reached often so the implementation
should be replaced. But this is a good start.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>