Extend CellActor with size for future
Supposedly in the future we may support actors composing multiple cells.
Implement the basis for this. Add getter and setter of "s" (size)
similarly to "c" (cell), and similarly sx and sy getters. By default the
size is (1, 1) and this is stored as None, since pretty much all actors
are of size 1x1.

The whole design of multi-cell actors is not trivial. Not for now.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>