Rename MOVE_TYPE constants and add new type MOVE_N for static objects
Prepare to adding cart objects that are similar to lifts, but move one
cell on floor rather than until the first obstacle on void.

New move type MOVE_N is intended for static (unmovable) lifts and carts.
This new type is not yet used, but may be used in the future.

Add new image for liftn in default theme that reuses lift-empty image.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>