Implement new mirror concept
Mirror is an object that reflects a beam (will soon be used in Mirror
Puzzle). It has orientation (one of four) and activeness (active or not,
plus flipped activeness or not). Inactive mirrors bypass a beam. Mirrors
with flipped activeness change their reflectiveness property depending on
whether it is the first or repeating encounter of the mirror by the beam.

Mirror is a property of cellactor that is None for regular actors and
contains an object of type Mirror inheriting from CellActor for mirror
hosts (carrier).

Currently the mirror carriers may be barrel or lift (and cart in near
future). Mirrors have separate image, but should be synced with their
carriers (position, movement, animation).

Update map chars accordingly (for regular actors and hosting a mirror).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>