Add some useful Area methods to query cell
Add these methods:
	* is_cell_evnevn cell
	* is_cell_evnodd cell
	* is_cell_oddevn cell
	* is_cell_oddodd cell
	* is_cell_inside cell [margin=0]
	* is_cell_on_margin cell [margin=0]

The first four answer whether the cell is in certain 2x2 grid subarea.
This is useful in some puzzles.

The is_cell_inside tells whether the given cell is inside area, excluding
an optional margin. Negative margin is supported too meaning a cell may
be slightly outside the area.

The is_cell_on_margin tells whether the given cell is on the margin
(False if margin is 0). Negative margin is supported too meaning a cell
may be slightly outside the area.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>