BarrelPuzzle: Generate only solvable rooms for zsb mode
Just shuffle the barrels maximally (so that the total distance is maximal
between the stock and the final barrel cells) using valid zsb barrel
moves only to get the final plate positions. This ensures the levels are
always solvable.

Still the puzzle may be easier than possible, since we don't reroll the
stock barrel cells (and they may dictate easier puzzles) for fast
generation, but this is a feature; the player may be blessed and receive
an easier puzzle. :)

Use the same get_all_valid_zsb_barrel_moves method for both generation
and solving. It returns pairs of (barrel_cell, target_cell) with distance
of 2 between the two. Additional get_all_valid_zsb_barrel_pushes method
converts this list to pairs of (char_cell, barrel_cells) with distance of
1 between the two for the zsb solver.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>