images: Add glass images to different themes
The glass images in minecraft and classic themes are unique.

In all other themes make variation of the same blue-cyan glass image,
just one non-transparent, the default that is slightly transparent and
the one that is dim.

The images were created in GIMP with some advice from the web and are
freely distributable as Free Art.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>