Add support for new cell types "trap0" and "trap1"
Traps are somewhat similar to gates in that they can block the path and
should be disarmed to free the path.

The meaning of gate states and trap states is reversed:
	* gate1 is open, but trap1 is a fatal hazard
	* gate0 is closed, but trap0 is disarmed (free to pass)

Unlike gate0, trap1 is not an obstacle. However stepping on it makes you
die and lose the level.

Use the usual has_trap mechanism in Puzzle to conditionally load them.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>