GatePuzzle: Add support for traps as well as gates
Everything already works with bits, so only minimal code should be added
to support traps as an alternative to (or together with) gates.

When "use_traps" is set in config, the traps are generated instead of
gates.

The maps to may now have a mix of gates and traps and the plates will
toggle either of them.

Continue to call everything "gates" (for example, num_gates and
toggle_gate) although now everything supports traps as well.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>