images: Add beam, beamgn and beamcl images to different themes
There is only one beam for now. It should look nice in most cases.

There are 3 variations of beam generator and collector.

The generator looks like a 4 pointed star, the collector is a circle or
rounded square.

The minecraft variation is made to be distinct, with beacon as a
generator and ender pearl as a collector.

The images were created in GIMP with some advice from the web and are
freely distributable as Free Art.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>