Puzzle: Add callback on_draw_map and simplify on_draw
The on_draw hook is still called at the end of the main draw, but before
the flash messages about end of game and level info. Do not pass mode
argument, since it is not related to draw. If it will ever be needed for
the puzzles, pass it using a different mechanism.

Add new callback on_draw_map that is called earlier than on_draw, just
after the map is drawn.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>