Add optional support for Pygame-Console
Ctrl-Escape now tries to create pgconsole. If this module is installed
(together with Pygame-BitmapFont) or symlinks to the corresponding source
directories pgconsole and pgbitmapfont are present, then the console is
shown in the top part of the game window. Press Escape to close it.

The console feature is added to Game singleton that will get more
features moved from main with a time.

pgzero does not allow a user to have pygame events, so there is a need to
hack the PGZeroGame handlers. When the console is enabled these pgzero
handlers are replaced with our own handlers, and they are restored when
the console is disabled:
	pygame.KEYDOWN
	pygame.KEYUP
	pygame.MOUSEBUTTONDOWN
	pygame.MOUSEBUTTONUP
	pygame.TEXTINPUT

Use gnu-free fonts for the console.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>