solution: Create solution string and copy it to clipboard
The solution string consists of direction letter when applicable, like
'r' for move to the right and 'R' for push to the right.

There are also these item substrings:
	"press(x, y)" or "press(x, y)^b" (when button number is not None)
	"goto(x, y)"
	"(x, y)" in bad cases when direction of move/push can't be resolved

The solution string consists of item substrings joined by a space.

The solution string is now silently copied to clipboard. Maybe some
status notification about this copying will be done in the future.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>