Call puzzle.on_prepare_enter_cell earlier
A puzzle may want to know when the character collects drops or opens
locks before this happends. So call puzzle.on_prepare_enter_cell at the
start of prepare_enter_cell and not at the end, to have a chance to catch
these events. The puzzle.is_char_phased callback is still called at the
end of prepare_enter_cell.

The puzzle may only catch drop and lock-open, but not prevent them, since
this would break the game mechanics, i.e. it is fine to read the current
map cell in this callback, but not modify it.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>