cmdargs: Add initial support for command line arguments
For now support just --help and --debug as prove of concept. More options
will be added as needed.

The arguments are added via environment variable CMD_ARGS, since there is
no other way to do it properly in pgzero.

Support of CMD_ARGS in dungeon.bat will be added later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>