GatePuzzle: Fix sounds during solution
The press_cell callback is always expected to return boolean whether the
press was handled. Not returning True on plate presses during solution
produces redundant error sound. Fix this omission.

Do the same fix in MinotaurPuzzle (it should always return True) although
press_cell is not currently used in MinotaurPuzzle solution, so there is
no sound problem in it.

Fixes: 2618868ad0f2 ("Handle Space centrally to mean press char cell rather than in each puzzle")
Signed-off-by: Mikhael Goikhman <migo@freeshell.org>