Redesign collections and levels
This is a complete redesign of the concept of collections and levels.

All levels are now contained in the corresponding collection.

Collection is an object with id, name, icon, puzzle-type and magic-n.
Collection also has level_configs, ordered configurations of levels

Collection id can be either string that is the same as the sub-path in
levels/ dir, like sokoban/original or gate or sokoban/dh/2. Or numeric
id, similar to magic-n, but with own logic to avoid duplicates.

Both collection ids (stringy and numeric) are unique. I.e. "gate" or "2".

Level id is just its collection id and its index number starting from 1
and delimited by a dot. Like "gate.4" or "2.4" - both are the same level.

Now all collections and their level configs are auto-registered at the
start, the directory levels/ is auto-scanned recursivelly for config
files. These are collection configs that contain either key levels or key
sokoban-map-file or sokoban-map-files.

In the case of sokoban-map-file or sokoban-map-files, the level
configurations are auto-generated from corresponding sokoban files as
usual. In the case of sokoban-map-files there is also a split into
sub-collections with sub-ids that are either indexes (like "sokoban/dh/1"
and sokoban/dh/2") if it is list or sub-names (like
"sokoban/micro/nabokosmos" and "sokoban/micro/sokogen") if it is dict.

There are new classes Collection and Level.

There is a singleton game.level and game.collections (all flat ordered
collections).

The Game class was enhanced with all level and collection navigation and
query methods, so now leveltools is redundant.

All places that accessed level as dict are updated to access object data.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>