Add reverse-barrel-mode and support it in BarrelPuzzle
In the reverse-barrel-mode the barrels and plates are swapped
and the barrel interactions are pulls instead of pushes.

Currently load_map and generate_room support this mode in BarrelPuzzle.
The meaning of barrels and plates are swapped. The character position is
currently the same as in the original level although it may cause level
to be unsolvable. In the future this will be improved.

Added one more level to sokoban/zsb with reverse-barrel-mode.

Currently solution does not work correctly in this mode (the push based
solution is found) that does not play correctly. But soon the pull based
solution will be found in this mode.

There is support for reverse-barrel-mode setting in collection, it is
used as the level default if absent. Soon a command line flag will be
supported as well.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>