Call the first enter_room after puzzle.store_level, not before it
This is a minor desigh error that was never visible until now, since
on_enter_room in puzzles never changed the level data being stored.
But for proper implementing reverse-barrel-mode toggling logic in
BarrelPuzzle we need to fix this flaw.

Continue to start music just after enter_room, as it always was.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>