BarrelPuzzle: Implement toggling reverse-barrel-mode and keep maps normal
Now the maps are always loaded, generated and reloaded in normal mode.
They are reversed in BarrelPuzzle only on reverse-barrel-mode at
on_enter_room.

Now Alt-E in BarrelPuzzle toggles reverse-barrel-mode in the level and
reloads the level similarly to Alt-R (that does not change
reverse-barrel-mode).

The changed reverse-barrel-mode in current level remains until the level
change (a win, or 'p', or 'n', or 'r'), then it gets reverse-barrel-mode
from scratch (from the level or colleciton config, or from command line).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>