Support special argument type letslogic:<ll-collection-id> for playing
This is somewhat similar to clipboard: argument, but instead of getting
sokoban levels from clipboard, fetch them from LetsLogic web site.

The loaded ll collections (and their levels) are cached locally, so there
will be no need to fetch them again.

Try:
	dungeon letslogic:999

All kinds of errors are hundled properly. It does not work without
internet obviously, and can be stuck if the web site is firewalled.

Implement the letslogic API and use puzzle-dungeon user for this.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>