debug: Refactor module functions and vars into object, implement features
Now the debug interface looks similar, but different.

debug(lvl, msg) still works as usual.

But now it is a singleton debug of class Debug behaving as a function
(via __call__ method).

There are now debug.lvl and debug.features. As well as debug.configure().

Implement the concept of debug features. Use one DBG_SOLV ("solv")
feature in BarrelPuzzle and GatePuzzle.

To enable features, now --debug SELECTOR is repeatable; the value may be
either int (debug level as usual) or any number of features, like "solv"
(meaning debug solving/solvers).

The debug() usage is now:

	debug("message")           print unconditionally
	debug(3, "message")        print if --debug LVL is greater than 3
	debug([2], 3, "message")   print with depth 2 (2 spaces prepended)
	debug([2], DBG_SOLV, "…")  print with --debug solv only; prepended
	debug([2], 3, DBG_SOLV, "…") printed with --debug 5 --debug solv

The order of the selectors can be any, but the last argument is always
the message.

The selector logic is to use AND between lvl and feature-condition, and
to use OR between features if multiple features are specified in debug().
I.e. any feature enabled via command line causes printing in this case.

Replace some debug(0) in puzzles with warn were it is a logical warning.

More debug features (other than "solv") may be supported later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>