Use find_map_file in load_map
For stock level configs this may be less needed, but when loading maps
from command line, the method find_map_file should be used, to support
maps with absolute filenames and maps relative to the current directory,
and not just to the data/maps directory.

Also handle properly the missing map file case as was intended, without
raising error.

Currently the internal map handling is different from parsing the sokoban
maps in how file open errors are processed. For internal maps, errors are
just ignored allowing Puzzle to possibly generate a level instead of
loading.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>