Handle single Shift, Alt or Ctrl press centrally, reserve mods for puzzles
Single shift or alt or ctrl generate on_key_down in pgzero. We handle it
in mainscreen, but this is better to be handled centrally. Just ignore
single modifier presses.

Also, reserve right Ctrl and/or right Alt together with regular keys to
the puzzles, don't handle such presses in main.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>