common, debug: Add get_current_time_str and use it in debug on request
New common funcion get_current_time_str returns the current day's time
string, like "17:20:14". Or "17:20:14.642345" if argument 6 is given.
Argument 0 is the default, but also 1 to 6 [meaning number of floating
digits] are supported.

Debug module now supports new show_time feature with the default of 3
floating digits. This is activated by --debug time. Or --debug time=6 to
also change the number of floating digits from the default 3 to 0..6.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>