BarrelPuzzle: Completely Rework Sokoban solver
Now 4 algorithms are supported:
	* IDDFS
	* BFS
	* PQ
	* A*

Currently IDDFS is implemented using recursive calls, just like previous
find_solution implementation.

All 4 algorithms support yielding to the GUI mainloop every second. So
now the solution progress is properly and timely reported.

Also, there is support for push optimal solution and move optional
solution now. KP-Enter now finds push-optimal and Shift-KP-Enter finds
move-optimal solution. Previously the first found solution was returned,
even non optimal. This required a redesign of data structures.

There is a completely new design for search nodes. Two classes Position
and SuperPosition are added. SuperPosition represents all common data
with the given barrel_cells. And Position is char entry in SuperPosition.
The actual search nodes are positions. Each position has parent position
and corresponding superposition. Each position may be expanded (childred
positions created) once. And may be reparented if a better path found to
that position.

This design was chosen for efficiency and performance reasons.

The progress is reported according to the algorithm. Some algorithms
(IDDFS, BFS) have depth limit that grows if the optimal solution is not
found. BFS grows by 1, and IDDFS grows by 5 (was 4). This may change.

Some algorithms have unprocessed_positions queue (PQ, A*) or list (BFS).
The difference between PQ and A* is in sort-positions cost function. For
PQ it is total_nums: (num_moves, num_shifts) or (num_shifts, num_moves).
For A* it is own solution_cost heuristics that combines total_nums with
the remaining minimal solution length from barrels to the nearest goals.

All should work with normal sokoban maps, special SZB maps, and
supposedly for reverse verstions of these.

The default algorithm is currently BFS, but A* is good too.

To change the algorithm press RAlt-A, RAlt-B, RAlt-D or RAlt-P.
Also RAlt-0 and RAlt-- are supported, mostly for debug purposes,
to disable_budget_solution and disable_prepare_solution.

It is possible to cancel solution earlier by pressing any key. In this
case the current best found solution is returned (if any), even if non
optimal. This is good for levels where it is quick to find a solution by
IDDFS or A*, but it is difficult to prove that it is the optimal (need to
process lots of positions for this). Although all positions that do not
improve the currently found solution are cut.

There were also other small improvements and some fixes on the way.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>